requirements.txt
ipdb==0.13.13
matplotlib==3.6.2
numpy==1.23.4
pandas==1.5.3
Pillow==9.3.0
PyYAML==6.0
scipy==1.9.1
timm==0.6.13
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.29.0
